Explicitly set contour zorder level above scatter
Previously the contour plot's zorder was not set and I found that it was
often, and randomly, below my scatter plot's zorder. Thus the plotted
points were hiding the contour plot. This is accomplished by
introducing a default_contour_args dictionary.

NOTE: this means that the point masking is not working!