GMMBayes updated from sklearn GMM to GaussianMixture
Updating Gaussian Mixture Model classifier to be concurent with the new
sklearn implementation GaussianMixture, appart from removing a barrage
of warning messages, solves the problem of the sklearn GMM class
getting depracated as of sklearn version 0.2.

GaussianMixture does not correspond one-to-one with old GMM implementation.
Forseable consequeces for code built on GMMBayes are:
    - min_covar is replaced by reg_covar (instantiation)

    - default GMM covariance_type is "full" instead of "diag"

    - GMM.covars_ is replaced by GaussianMixture.covariances_
      therefore .gmms_[].covars_ becomes .gmms_[].covariances_[]

    - GMM.score is replaced by GaussianMixture.score_samples

    - GaussianMixture implementation limits the number of gaussian
      components to number of sample points they are fitted to. This
      did not seem to be the case with GMM therefore additional check
      is implemented and number of gaussian components of the model
      is now manually set to number of sample points if the number
      of gaussian components was larger than the number of sample
      points being fitted. A warning message is issued.

The results from GaussianMixture seem, on surface level, to correspond
to results from GMM. Further testing desired.